WIP: add UntypedLiteral op based on type hints in Ops
I think this might be a dead end. This works by looking at the type hints of ops, and at expression construction time intercepting the `__coerce__` logic and determining if the op declares that it prefers untyped literals.

But the flaw is shown if you consider there two cases:
`t.x > ibis.literal(12.34)` should compile to `x > 12.34`, since whether or not duckdb interprets the 12.34 as a float64 or a float32 or a decimal doesn't really matter.
`t.select(col=ibis.literal(12.34))` should compile to `12:34::float64 as col`, since here the resulting datatype is very important, so we need to tell the backedn explicitly what type we want.

Note that in both cases, there is a `ibis.literal(12.34)` in the expression tree. In one instance, it wants a cast, in the other, it doesn't. But we can only know which one we want by considering the larger context of where this expression is: if it is inside a comparison, then drop the cast, otherwise we want it. So, I think this NEEDs to be implemented as rewrite rules that work over the entire expression tree.